update: admin master pages remove edit button
- Now edit include in details page,
- Details pages now show form like create with initial value of what subject detailed
