+++ resources/js/Components/PegawaiFormDataOrganisasi.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, Dispatch, memo, SetStateAction } from "react";
 import { Button, Card, Tooltip, Typography } from "@material-tailwind/react";
 import { Input } from "@/Components/Input";
 import { ListPlus, ListX } from "lucide-react";
-import type { FormDataOrganisasi } from "@/Pages/Admin/ADMIN_PegawaiCreatePage";
+import type { FormDataOrganisasi } from "@/Pages/Admin/MASTER_PegawaiCreatePage";
 
 const PegawaiFormDataOrganisasi = ({ formState, setFormState, formInitial }: {
     formState: FormDataOrganisasi[];
