+++ resources/js/Components/PegawaiFormDataPendidikanFormal.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, Dispatch, memo, SetStateAction } from "react";
 import { Button, Card, Option, Select, Tooltip, Typography } from "@material-tailwind/react";
 import { Input } from "@/Components/Input";
 import { ListPlus, ListX } from "lucide-react";
-import type { FormDataPendidikanFormal } from "@/Pages/Admin/ADMIN_PegawaiCreatePage";
+import type { FormDataPendidikanFormal } from "@/Pages/Admin/MASTER_PegawaiCreatePage";
 
 const PegawaiFormDataPendidikanFormal = ({ formState, setFormState, formInitial }: {
     formState: FormDataPendidikanFormal[];
