+++ resources/js/Pages/Admin/MASTER_UnitIndexPage.tsx
@@ -9,55 +9,88 @@ import {
     Typography
 } from "@material-tailwind/react";
 import {
-    ArrowLeft, ArrowRight, Award,
-    ChevronDown, FileSearch,
-    Pencil,
+    ArrowLeft, ArrowRight,
+    BarChartBig,
+    Building2,
+    ChevronDown, Download, FileSearch,
+    NotebookText,
     Plus,
     Search, Trash2,
+    User2
 } from "lucide-react";
 import { MTColor, PageProps } from "@/types";
 import { AdminLayout } from "@/Layouts/AdminLayout";
-import { Head, router } from "@inertiajs/react";
+import { Head, Link, router } from "@inertiajs/react";
 import { Input } from "@/Components/Input";
 import { format } from "date-fns";
 import { id as localeID } from "date-fns/locale/id";
 import { TextArea } from "@/Components/TextArea";
+import { Checkbox } from "@/Components/Checkbox";
 import { useTheme } from "@/Hooks/useTheme";
 import { SyntheticEvent, useEffect, useMemo, useState } from "react";
 import { Bounce, toast } from "react-toastify";
 import {  z } from "zod";
 import axios, { AxiosError } from "axios";
+import * as XLSX from "xlsx";
+import { id } from "date-fns/locale";
 
-export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<{
-    marhalahs: {
+export default function MasterManageUnitPage({ auth, units, adminCount }: PageProps<{
+    units: {
         id: string;
         nama: string;
         keterangan: string;
         created_at: string;
+        admin: {
+            id: string;
+            username: string;
+            unit_id: string;
+        }[] | []
     }[] | [];
+    adminCount: number;
 }>) {
-    const TABLE_HEAD = ['No', 'Nama', 'Keterangan', 'Tanggal daftar', 'Aksi'];
+    const TABLE_HEAD = ['No', 'Nama', 'Admin', 'Tanggal daftar', 'Aksi'];
     const cardData = [
         {
             color: "gray",
-            icon: <Award />,
-            title: "Jumlah Marhalah terdaftar",
-            value: marhalahs.length,
-        }
+            icon: <Building2 />,
+            title: "Jumlah Unit terdaftar",
+            value: units.length,
+        },
+        {
+            color: "gray",
+            icon: <User2 />,
+            title: "Jumlah Akun Unit aktif",
+            value: adminCount,
+        },
+        {
+            color: "gray",
+            icon: <NotebookText />,
+            title: "Pengajuan Promosi Unit",
+            value: "1",
+        },
+        {
+            color: "gray",
+            icon: <BarChartBig />,
+            title: "Lorem ipsum",
+            value: "704",
+        },
     ];
+    const deleteDialogInit = {
+        open: false,
+        unitId: '',
+        admins: []
+    };
     const { theme } = useTheme();
     const [ openFormDialog, setOpenFormDialog ] = useState(false);
     const [ deleteDialog, setDeleteDialog ] = useState<{
         open: boolean;
-        id: string;
-        nama: string;
-        onSubmit: boolean;
-    }>({
-        open: false,
-        id: '',
-        nama: '',
-        onSubmit: false
-    });
+        unitId: string;
+        admins: {
+            id: string;
+            username: string;
+            unit_id: string;
+        }[]
+    }>(deleteDialogInit);
     const [ sortBy, setSortBy ] = useState('');
     const [ currPage, setCurrPage ] = useState(1);
     const [ viewPerPage, setViewPerPage ] = useState(10);
@@ -78,10 +111,10 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
         const startIndex = (currPage - 1) * viewPerPage;
         const lastIndex = startIndex + viewPerPage;
 
-        return marhalahs.slice(startIndex, lastIndex);
-    }, [ marhalahs, viewPerPage ]);
+        return units.slice(startIndex, lastIndex);
+    }, [ units, viewPerPage ]);
 
-    const [ data, setData ] = useState(adjustData);
+    const [data, setData] = useState(adjustData);
     const [ search, setSearch ] = useState('');
     const getItemProps = (index: number) =>
         ({
@@ -91,7 +124,7 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
         } as any);
 
     const nextPage = () => {
-        const totalPages = Math.ceil(marhalahs.length / viewPerPage);
+        const totalPages = Math.ceil(units.length / viewPerPage);
         currPage < totalPages && setCurrPage(currPage + 1);
     };
     const prevPage = () => {
@@ -122,6 +155,30 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
         onSubmit: boolean;
     }>(formInputInit);
 
+    const handleXLSXDownload = () => {
+        const workbook = XLSX.utils.book_new();
+        const data = units.map((unit) => ({
+            nama: unit.nama,
+            keterangan: unit.keterangan,
+            tanggal_daftar: format(unit.created_at, 'PPPP', {
+                locale: id
+            })
+        }))
+        const sheet = XLSX.utils.json_to_sheet(data, {
+            header: ['nama', 'keterangan', 'tanggal_daftar'],
+        });
+        XLSX.utils.book_append_sheet(workbook, sheet, 'Sheet1');
+
+        const wbout = XLSX.write(workbook, { bookType: 'xlsx', type: 'array' });
+        const blob = new Blob([wbout], { type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet' });
+        const url = URL.createObjectURL(blob);
+        const link = document.createElement('a');
+        link.href = url;
+        link.download = 'data.xlsx';
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+    };
     const formSubmitDisabled = (): boolean => (!formInput.nama || !formInput.keterangan);
     const handleFormSubmit = (event: SyntheticEvent<HTMLFormElement>) => {
         event.preventDefault();
@@ -131,11 +188,11 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
             ...prevState,
             onSubmit: true
         }));
-        const marhalahSchema = z.object({
-            nama: z.string().min(1, { message: "Nama Marhalah tidak boleh kosong" }),
-            keterangan: z.string().min(1, { message: "Keterangan Marhalah tidak boleh kosong" })
+        const unitSchema = z.object({
+            nama: z.string().min(1, { message: "Nama Unit tidak boleh kosong" }),
+            keterangan: z.string().min(1, { message: "Keterangan Unit tidak boleh kosong" })
         });
-        const zodUnitResult = marhalahSchema.safeParse({
+        const zodUnitResult = unitSchema.safeParse({
             nama: nama,
             keterangan: keterangan
         });
@@ -144,60 +201,45 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
             notifyToast('error', errorMessages, theme as 'light' | 'dark');
         }
 
-        axios.post(route('marhalah.create'), {
+        axios.post(route('unit.create'), {
             nama: nama,
-            keterangan: keterangan
+            keterangan: keterangan,
         })
             .then(() => {
-                notifyToast('success', 'Marhalah berhasil ditambahkan!', theme as 'light' | 'dark');
-                router.reload({ only: ['marhalahs'] });
+                notifyToast('success', 'Unit berhasil ditambahkan!', theme as 'light' | 'dark');
+                router.reload({ only: ['units'] });
+                setOpenFormDialog(false);
             })
             .catch((err: unknown) => {
-                const errMsg: string = err instanceof AxiosError
-                    ? err.response?.data.message ?? 'Error tidak diketahui terjadi!'
+                const errMsg = err instanceof AxiosError
+                    ? err.response?.data.message as string ?? 'Error tidak diketahui terjadi!'
                     : 'Error tidak diketahui terjadi!'
                 notifyToast('error', errMsg, theme as 'light' | 'dark');
             })
             .finally(() => {
-                setOpenFormDialog(false);
                 setFormInput((prevState) => ({ ...prevState, onSubmit: false }))
             });
     };
-    const handleDeleteSubmit = () => {
-        const { id } = deleteDialog
-
-        setDeleteDialog((prevState) => ({
-            ...prevState,
-            onSubmit: true
-        }));
-        const marhalahschema = z.object({
-            id: z.string().min(1, { message: "Marhalah belum terpilih" }),
-        });
-        const zodUnitResult = marhalahschema.safeParse({
-            id: id,
-        });
-        if (!zodUnitResult.success) {
-            const errorMessages = zodUnitResult.error.issues[0].message;
-            notifyToast('error', errorMessages, theme as 'light' | 'dark');
-        }
-
-        axios.post(route('marhalah.delete'), {
-            id: id,
+    const handleDeleteUnit = () => {
+        axios.post(route('unit.delete'), {
+            id: deleteDialog.unitId,
+            admins: deleteDialog.admins.length > 0
+                ? deleteDialog.admins.map((admin) => admin.id)
+                : []
         })
             .then(() => {
-                notifyToast('success', 'Marhalah berhasil dihapus!', theme as 'light' | 'dark');
-                router.reload({ only: ['marhalahs'] });
+                notifyToast('success', 'Unit terpilih berhasil dihapus!');
+                setData((prevState) => prevState.filter((filt) => filt.id !== deleteDialog.unitId));
+                setDeleteDialog(deleteDialogInit);
             })
             .catch((err: unknown) => {
-                const errMsg: string = err instanceof AxiosError
-                    ? err.response?.data.message ?? 'Error tidak diketahui terjadi!'
-                    : 'Error tidak diketahui terjadi!'
-                notifyToast('error', errMsg, theme as 'light' | 'dark');
-            })
-            .finally(() => {
-                setDeleteDialog((prevState) => ({ ...prevState, onSubmit: false, open: false }));
+                const errMsg = err instanceof AxiosError
+                    ? err?.response?.data.message ?? 'Error tidak diketahui terjadi'
+                    : 'Error tidak diketahui terjadi';
+                notifyToast('error', errMsg);
             });
     };
+
     useEffect(() => {
         if (!openFormDialog) {
             setFormInput(formInputInit);
@@ -205,20 +247,23 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
     }, [ openFormDialog ]);
     useEffect(() => {
         setData(adjustData);
-    }, [ marhalahs, viewPerPage ]);
+    }, [ units, viewPerPage ]);
     useEffect(() => {
         if (search.length < 1) {
             setData(adjustData);
         } else {
             setCurrPage(1);
-            const matchmarhalahs = marhalahs.filter((marhalah) => marhalah.nama.toLowerCase().includes(search.toLowerCase()));
-            setData(matchmarhalahs);
+            const matchUnits = units.filter(unit =>
+                unit.nama.toLowerCase().includes(search.toLowerCase()) ||
+                unit.admin.some(admin => admin.username.toLowerCase().includes(search.toLowerCase()))
+            );
+            setData(matchUnits);
         }
     }, [ search ]);
 
     return (
         <>
-            <Head title="Master - Marhalah" />
+            <Head title="Master - Unit" />
             <AdminLayout>
                 <section className="mb-1 grid gap-y-10 gap-x-6 md:grid-cols-2 xl:grid-cols-4">
                     { cardData.map(({ icon, title, color, value }) => (
@@ -249,19 +294,26 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                         <div className="mb-8 flex items-center justify-between gap-x-3">
                             <div>
                                 <Typography variant="h5" color="blue-gray">
-                                    Daftar Marhalah
+                                    Daftar Unit
                                 </Typography>
                                 <Typography color="gray" className="mt-1 font-normal">
-                                    Informasi mengenai Marhalah yang terdaftar
+                                    Informasi mengenai Unit yang terdaftar
                                 </Typography>
                             </div>
-                            <div className="flex shrink-0 flex-col gap-2 sm:flex-row">
+                            <div className="flex flex-col shrink-0 gap-2 lg:flex-row">
                                 <Button
                                     onClick={() => setOpenFormDialog(true)}
                                     className="flex items-center gap-1.5 capitalize font-medium text-base" size="sm"
                                 >
                                     <Plus />
-                                    Tambahkan Marhalah baru
+                                    Tambahkan Unit baru
+                                </Button>
+                                <Button
+                                    onClick={handleXLSXDownload}
+                                    className="flex items-center gap-1.5 capitalize font-medium text-base" size="sm"
+                                >
+                                    <Download />
+                                    Unduh Data (XLSX)
                                 </Button>
                             </div>
                         </div>
@@ -306,7 +358,7 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                             <tbody>
                             {
                                 data.map(
-                                    ({ id, nama, keterangan, created_at }, index) => {
+                                    ({ id, nama, admin, created_at }, index) => {
                                         const isLast = index === data.length - 1;
                                         const classes = isLast
                                             ? "p-4"
@@ -325,6 +377,7 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                                                 </td>
                                                 <td className={ `${ classes } min-w-52` }>
                                                     <div className="flex items-center gap-3">
+                                                        {/*<Avatar src={img} alt={name} size="sm" />*/ }
                                                         <div className="flex flex-col">
                                                             <Typography
                                                                 variant="small"
@@ -337,16 +390,26 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                                                     </div>
                                                 </td>
                                                 <td className={ `${ classes } min-w-52` }>
-                                                    <div className="flex items-center gap-3">
-                                                        <div className="flex flex-col">
-                                                            <Typography
-                                                                variant="small"
-                                                                color="blue-gray"
-                                                                className="font-normal"
-                                                            >
-                                                                { keterangan }
-                                                            </Typography>
-                                                        </div>
+                                                    <div className="flex flex-col gap-1">
+                                                        {
+                                                            admin.length < 1
+                                                                ? (
+                                                                    <div
+                                                                        className="flex items-center justify-start text-xs text-gray-400">
+                                                                        Belum ada Admin untuk unit ini
+                                                                    </div>
+                                                                )
+                                                                : admin.map((admn, index) => ((
+                                                                    <Link
+                                                                        href={route('master.admin.details')}
+                                                                        data={{ q: admn.id }}
+                                                                        key={ index }
+                                                                        className="font-normal text-sm hover:text-blue-600"
+                                                                    >
+                                                                        { admn.username }
+                                                                    </Link>
+                                                                )))
+                                                        }
                                                     </div>
                                                 </td>
                                                 <td className={ `${ classes } min-w-40` }>
@@ -363,14 +426,11 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                                                 <td className={ classes }>
                                                     <div className="w-32 flex gap-2.5 items-center justify-start">
                                                         <Tooltip content="Detail">
-                                                            <IconButton variant="text">
-                                                                <FileSearch className="h-5 w-5 text-blue-800"/>
-                                                            </IconButton>
-                                                        </Tooltip>
-                                                        <Tooltip content="Edit">
-                                                            <IconButton variant="text">
-                                                                <Pencil className="h-5 w-5"/>
-                                                            </IconButton>
+                                                            <Link href={route('master.unit.details', { q: units[index].id })}>
+                                                                <IconButton variant="text">
+                                                                    <FileSearch className="h-5 w-5 text-blue-800"/>
+                                                                </IconButton>
+                                                            </Link>
                                                         </Tooltip>
                                                         <Tooltip content="Hapus" className="bg-red-400">
                                                             <IconButton
@@ -379,8 +439,8 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                                                                     setDeleteDialog((prevState) => ({
                                                                         ...prevState,
                                                                         open: true,
-                                                                        id: id,
-                                                                        nama: marhalahs.find((marhalah) => marhalah.id === id)?.nama ?? '-'
+                                                                        unitId: id,
+                                                                        admins: admin
                                                                     }))
                                                                 }}
                                                             >
@@ -445,10 +505,10 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                         <Card>
                             <CardBody className="flex flex-col gap-4">
                                 <Typography variant="h4" color="blue-gray">
-                                    Menambahkan Marhalah baru
+                                    Menambahkan Unit baru
                                 </Typography>
                                 <Input
-                                    label="Nama Marhalah"
+                                    label="Nama unit"
                                     size="lg"
                                     required={true}
                                     error={formInput.error.nama}
@@ -481,6 +541,28 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                                         }));
                                     }}
                                 />
+                                {/*<Checkbox*/}
+                                {/*    checked={formInput.isMaster}*/}
+                                {/*    onChange={() => {*/}
+                                {/*        setFormInput((prevState) => ({*/}
+                                {/*            ...prevState, isMaster: !prevState.isMaster*/}
+                                {/*        }));*/}
+                                {/*    }}*/}
+                                {/*    label={*/}
+                                {/*        <Typography color="blue-gray" className="flex font-medium">*/}
+                                {/*            Berikan Hak Master. Apa itu Hak master?*/}
+                                {/*            <Typography*/}
+                                {/*                as="a"*/}
+                                {/*                href="#"*/}
+                                {/*                color="blue"*/}
+                                {/*                className="font-medium transition-colors hover:text-blue-700"*/}
+                                {/*            >*/}
+                                {/*                &nbsp;Pelajari selengkapnya*/}
+                                {/*            </Typography>*/}
+                                {/*            .*/}
+                                {/*        </Typography>*/}
+                                {/*    }*/}
+                                {/*/>*/}
                             </CardBody>
                             <CardFooter className="pt-0 flex gap-3 justify-between">
                                 <Button
@@ -503,34 +585,47 @@ export default function MasterManageMarhalahPage({ auth, marhalahs }: PageProps<
                         </Card>
                     </form>
                 </Dialog>
+
                 <Dialog open={deleteDialog.open} handler={handleOpenDelete}>
                     <DialogHeader className="text-gray-900">
-                        Hapus Marhalah terpilih?
+                        Hapus Unit terpilih?
                     </DialogHeader>
                     <DialogBody>
                         <Typography variant="h6" className="text-gray-900 truncate">
                             Anda akan menghapus
-                            Marhalah: &nbsp;
-                            <span className="font-semibold">
-                                " { deleteDialog.nama } "
+                            Unit:&nbsp;
+                            <span className="font-bold">
+                                " { units.find((unit) => unit.id === deleteDialog.unitId)?.nama ?? '-' } "
                             </span>
                         </Typography>
-                        <p className="text-sm text-gray-900 font-medium">
+                        <p className="mt-1.5 text-sm text-gray-900 font-medium">
                                 <span className="text-red-600 font-bold">
                                     *
                                 </span>
-                            Pegawai yang terhubung akan akan kehilangan status marhalah
+                            Admin yang terhubung dengan Unit akan ikut dihapus:
                         </p>
+                        <ul className="flex flex-col gap-1.5 h-40 overflow-auto p-2 border-2 text-gray-800 font-medium">
+                            {
+                                deleteDialog.admins.map((admin) => ((
+                                    <li key={admin.id}>
+                                        <span>-</span> { admin.username }
+                                    </li>
+                                )))
+                            }
+                            {
+                                deleteDialog.admins.length < 1 && ( <p className="text-sm italic font-medium">Unit belum memiliki Admin</p>)
+                            }
+                        </ul>
                     </DialogBody>
                     <DialogFooter>
                         <Button
                             color="black"
-                            onClick={ () => setDeleteDialog((prevState) => ({ ...prevState, open: false })) }
+                            onClick={ () => setDeleteDialog(deleteDialogInit)}
                             className="mr-1"
                         >
-                            <span>Batal</span>
+                        <span>Batal</span>
                         </Button>
-                        <Button color="red" onClick={handleDeleteSubmit} loading={deleteDialog.onSubmit}>
+                        <Button color="red" onClick={handleDeleteUnit}>
                             <span>Hapus</span>
                         </Button>
                     </DialogFooter>
